% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_df.R
\name{boot_df}
\alias{boot_df}
\title{boot_df}
\usage{
boot_df(x_matrix, y_sur, ID = NULL, size = 10^5, seed = 1, prob = NULL)
}
\arguments{
\item{x_matrix}{matrix to perform sampling on}

\item{y_sur}{The numeric vector of the qualitative surrogate.}

\item{ID}{The patient ID}

\item{size}{size of matrix returned}

\item{seed}{seed for sampling}

\item{prob}{Vector for weight sampling}
}
\value{
A list with the sampled explanatory matrix and the sampled qualitative surrogate (y_sur)
}
\description{
Sample rows with replacement from a matrix
}
