% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_quali.R
\name{build_quali}
\alias{build_quali}
\title{build_quali}
\usage{
build_quali(x, p, q)
}
\arguments{
\item{x}{A numeric vector}

\item{p}{The lower quantile}

\item{q}{The upper quantile}
}
\value{
The qualitative surrogate (x in three categories) defining the extrema populations
}
\description{
build_quali
}
