% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg_test_phevis.R
\name{check_arg_test_phevis}
\alias{check_arg_test_phevis}
\title{check_arg_test_phevis}
\usage{
check_arg_test_phevis(
  train_param,
  df_test,
  surparam,
  model,
  START_DATE,
  PATIENT_NUM,
  ENCOUNTER_NUM
)
}
\arguments{
\item{train_param}{Parameters for the model training (variables used, main ICD and CUIS, half_life, gold standard, omega). Usually obtained from train_phevis() function.}

\item{df_test}{The dataframe on which to make the prediction.}

\item{surparam}{The parameters used to compute the surrogate. Usually obtained by train_phevis() function.}

\item{model}{The random intercept logistic regression. Usually obtained by train_phevis() function.}

\item{START_DATE}{Column name of the time column. The time column should be numeric}

\item{PATIENT_NUM}{Column name of the patient id column.}

\item{ENCOUNTER_NUM}{Column name of the encounter id column.}
}
\value{
No return value, stop the code execution if one condition is not met.
}
\description{
Function to check arguments passed to test_phevis()
}
