% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg_train_phevis.R
\name{check_arg_train_phevis}
\alias{check_arg_train_phevis}
\title{check_arg_train_phevis}
\usage{
check_arg_train_phevis(
  half_life,
  df,
  START_DATE,
  PATIENT_NUM,
  ENCOUNTER_NUM,
  var_vec,
  main_icd,
  main_cui,
  rf,
  p.noise,
  bool_SAFE,
  omega,
  GS
)
}
\arguments{
\item{half_life}{Duration of cumulation. For a chronic disease you might chose Inf, for acute disease you might chose the duration of the disease.}

\item{df}{\code{data.frame} containing all the variables.}

\item{START_DATE}{Column name of the time column. The time column should be numeric}

\item{PATIENT_NUM}{Column name of the patient id column.}

\item{ENCOUNTER_NUM}{Column name of the encounter id column.}

\item{var_vec}{Explanatory variables used for the prediction, including the main variables.}

\item{main_icd}{Character vector of the column names of the main ICD codes.}

\item{main_cui}{Character vector of the column names of the main CUIs.}

\item{rf}{should pseudo-labellisation with random forest be used (default is true)}

\item{p.noise}{percentage of noise introduced during the noising step (default is 0.3)}

\item{bool_SAFE}{A boolean. If TRUE, SAFE selection is done, else it is not (default is TRUE)}

\item{omega}{Constant for the extrema population definition (default is 2)}

\item{GS}{Character string corresponding to the name of the gold-standard variable (default is null for which a vector of 0 will be taken).}
}
\value{
No return value, stop the code execution if one condition is not met.
}
\description{
Function to check arguments passed to train_phevis()
}
