% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_lag.R
\name{cum_lag}
\alias{cum_lag}
\title{cum_lag}
\usage{
cum_lag(x, n_lag)
}
\arguments{
\item{x}{numeric vector for which lag variable should be computed}

\item{n_lag}{size of lag window}
}
\value{
return numeric vector.
}
\description{
helpful function to cumulate information.
}
