% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_perf.R
\docType{data}
\name{data_perf}
\alias{data_perf}
\title{Control data for test}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
data(data_perf)
}
\description{
Simulated dataset for PheVis phenotyping.
}
\keyword{datasets}
