% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phevis.R
\docType{data}
\name{data_phevis}
\alias{data_phevis}
\title{PheVis simulated dataset}
\format{
An object of class \code{data.frame} with 19659 rows and 15 columns.
}
\usage{
data(data_phevis)
}
\description{
Simulated dataset for PheVis phenotyping.
}
\keyword{datasets}
