% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expcorrectC}
\alias{expcorrectC}
\title{expcorrectC}
\usage{
expcorrectC(mat, diffdate, lambda)
}
\arguments{
\item{mat}{A matrix where each column is a variable to be cumulated.}

\item{diffdate}{Number of days between each sojourn. NA for switch of patient and restart cumulation.}

\item{lambda}{A double to set the exponential cumulation.}
}
\value{
A matrix corresponding to the mat argument with cumulated exponential decay
}
\description{
c++ function to compute exponential cumulation of information.
}
\details{
expcorrectC
}
