% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggindividual_plot.R
\name{ggindividual_plot}
\alias{ggindividual_plot}
\title{ggindividual_plot}
\usage{
ggindividual_plot(subject, time, gold_standard, prediction)
}
\arguments{
\item{subject}{numeric vector subject id}

\item{time}{numeric vector time or date}

\item{gold_standard}{numeric vector of gold standard}

\item{prediction}{numeric vector of prediction}
}
\value{
a ggplot graph
}
\description{
Plot individual predictions.
}
\examples{
ggindividual_plot(subject = rep(1,10),
  time = 1:10,
  gold_standard = c(0,0,1,1,0,0,1,1,0,0),
  prediction = runif(n = 10, min = 0, max = 1))

}
