% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_exp_smooth.R
\name{matrix_exp_smooth}
\alias{matrix_exp_smooth}
\title{matrix_exp_smooth}
\usage{
matrix_exp_smooth(half_life, df, date, patient_id, encounter_id)
}
\arguments{
\item{half_life}{Duration of accumulation. For a chronic disease you might chose Inf, for acute disease you might chose the duration of the disease.}

\item{df}{Dataframe of the explanatory variables.}

\item{date}{Vector of date. The date should be in a numeric format.}

\item{patient_id}{The vector of patient id}

\item{encounter_id}{The vector of visit id}
}
\value{
A data.frame object with both the raw variables and the accumulated ones.
}
\description{
Function to accumulate the information with exponential decay.
}
