% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noising.R
\name{noising}
\alias{noising}
\title{noising}
\usage{
noising(X_boot, p = 0.3)
}
\arguments{
\item{X_boot}{matrix to perform noise on}

\item{p}{amount of noise}
}
\value{
A noised matrix
}
\description{
Noise a matrix
}
