% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenorm_longit_simpl.R
\name{phenorm_longit_simpl}
\alias{phenorm_longit_simpl}
\title{phenorm_longit_simpl}
\usage{
phenorm_longit_simpl(
  df,
  var_surrogate,
  surrogates_quali,
  id_rnd,
  rf = FALSE,
  ntree = 100,
  bool_weight = FALSE,
  p.noise = 0.3,
  bool_SAFE = TRUE,
  size = 10^5
)
}
\arguments{
\item{df}{dataframe}

\item{var_surrogate}{variables used for building the surrogates}

\item{surrogates_quali}{numeric vector of the qualitative surrogate}

\item{id_rnd}{ID for random effect}

\item{rf}{should pseudo-labellisation with random forest be used (default is FALSE)}

\item{ntree}{number of tree for \code{randomforest} (default is 100)}

\item{bool_weight}{should the sampling probability balance the number of positive and negative extrema.}

\item{p.noise}{percentage of noise introduced during the noising step}

\item{bool_SAFE}{A boolean. If TRUE, SAFE selection is done, else it is not (default is TRUE)}

\item{size}{minimum size of sampling}
}
\value{
A list with the logistic model, the random forest model, the variables selected for prediction and the predictions
}
\description{
'PheNorm' like function adapted to longitudinal data.
}
