% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheVis-package.R
\docType{package}
\name{PheVis-package}
\alias{PheVis}
\alias{PheVis-package}
\title{PheVis: Automatic Phenotyping of Electronic Health Record at Visit Resolution}
\description{
Using Electronic Health Record (EHR) is difficult because most of the time the true characteristic of the patient is not available. Instead we can retrieve the International Classification of Disease code related to the disease of interest or we can count the occurrence of the Unified Medical Language System. None of them is the true phenotype which needs chart review to identify. However chart review is time consuming and costly. 'PheVis' is an algorithm which is phenotyping (i.e identify a characteristic) at the visit level in an unsupervised fashion. It can be used for chronic or acute diseases. An example of how to use 'PheVis' is available in the vignette. Basically there are two functions that are to be used: `train_phevis()` which trains the algorithm and `test_phevis()` which get the predicted probabilities. The detailed method is described in preprint by Ferté et al. (2020) \doi{10.1101/2020.06.15.20131458}.
}
\author{
\strong{Maintainer}: Thomas Ferte \email{thomas.ferte@u-bordeaux.fr}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
}

}
\keyword{internal}
