% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheVis.R
\docType{package}
\name{PheVis}
\alias{PheVis}
\title{PheVis: A package for visit phenotyping.}
\description{
Package implementing the PheVis algorithm for which a paper is in progress. The idea is to provide an unsupervised tool able to phenotype easily multiple phenotypes without chart review.
}
