% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_lme4model.R
\name{pred_lme4model}
\alias{pred_lme4model}
\title{pred_lme4model}
\usage{
pred_lme4model(model = NULL, fe.model = NULL, df)
}
\arguments{
\item{model}{lme4 model}

\item{fe.model}{the fixed effect of a model}

\item{df}{dataframe for prediction}
}
\value{
A vector of the predictions
}
\description{
function to predict probability from 'lme4' or 'glm' objects
}
