% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_cv.glmnet.R
\name{pretty_cv.glmnet}
\alias{pretty_cv.glmnet}
\title{pretty_cv.glmnet}
\usage{
pretty_cv.glmnet(
  x_glmnet,
  y,
  alpha = 1,
  family = "binomial",
  s = "lambda.1se",
  weights = rep(1, nrow(x_glmnet)),
  ...
)
}
\arguments{
\item{x_glmnet}{Independent variable matrix (X)}

\item{y}{Dependent variable vector (Y)}

\item{alpha}{alpha parameter of glmnet (default = 1)}

\item{family}{family parameter of glmnet (default = "binomial")}

\item{s}{lambda chosen from cv.glmnet (default = "lambda.1se")}

\item{weights}{glmnet parameter}

\item{...}{additional parameters passed to glmnet}
}
\value{
A list with the model, the coefficient associated with variables and the selected variables.
}
\description{
Train a 'glmnet' with cross validation (cv) model and return convenient results (model and results with non zero coefficients)
}
