% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_time_sum.R
\name{roll_time_sum}
\alias{roll_time_sum}
\title{roll_time_sum}
\usage{
roll_time_sum(
  id,
  id_encounter,
  var,
  start_date,
  win_size1 = 30,
  win_size2 = 365,
  name1 = "cum_month",
  name2 = "cum_year"
)
}
\arguments{
\item{id}{Patient id numeric vector}

\item{id_encounter}{Encounter id vector}

\item{var}{Variable numeric vector}

\item{start_date}{Time numeric vector}

\item{win_size1}{First window size (default is 30)}

\item{win_size2}{Second window size (default is 365)}

\item{name1}{name of first rolling var (default is "cum_month")}

\item{name2}{name of second rolling var (default is "cum_year")}
}
\value{
A dataframe containing the rolling variables.
}
\description{
Compute the cumulated information of what happened in past month and past year.
}
