% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_var.R
\name{rolling_var}
\alias{rolling_var}
\title{rolling_var}
\usage{
rolling_var(id, var, start_date, id_encounter)
}
\arguments{
\item{id}{Patient id numeric vector}

\item{var}{Variable numeric vector}

\item{start_date}{Time numeric vector}

\item{id_encounter}{Encounter id vector}
}
\value{
A dataframe containing the rolling variables.
}
\description{
Compute rolling variables (last visit, last 5 visits, last month and last year)
}
