% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_exp_smooth.R
\name{sur_exp_smooth}
\alias{sur_exp_smooth}
\title{sur_exp_smooth}
\usage{
sur_exp_smooth(half_life, sur, date, patient_id, encounter_id)
}
\arguments{
\item{half_life}{Duration of cumulation. For a chronic disease you might chose Inf, for acute disease you might chose the duration of the disease.}

\item{sur}{The quantitative surrogate.}

\item{date}{A numeric vector of time of days unit.}

\item{patient_id}{Vector of patient ID}

\item{encounter_id}{Vector of encounter ID}
}
\value{
A dataframe with the cumulated surrogate.
}
\description{
Function to cumulate surrogate with exponential decay
}
